backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
dnspython==2.6.1
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.7
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
keyring==23.5.0
langchain==0.2.11
langchain-core==0.2.24
langchain-ollama==0.1.0
langchain-text-splitters==0.2.2
langsmith==0.1.93
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
mongoengine==0.28.2
mongomock==4.1.2
more-itertools==8.10.0
multidict==6.0.5
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
ollama==0.3.0
orjson==3.10.6
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
PyGObject==3.42.1
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==2.4.7
pytest==8.3.2
pytest-mock==3.14.0
PyYAML==6.0.1
requests==2.32.3
SecretStorage==3.3.1
sentinels==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
systemd-python==234
tenacity==8.5.0
tomli==2.0.1
urllib3==2.2.2
wadllib==1.3.6
Werkzeug==3.0.3
yarl==1.9.4
zipp==1.0.0
